all PR updates: organization/naming/abstractions
