+++ src/components/Listings/Listings.js
@@ -1,7 +1,15 @@
 import React from 'react';
-import ListingsTabs from './ListingsTabs/ListingsTabs';
+import { Col, Container, Form } from 'reactstrap';
+import ListingsTabs from "./ListingsTabs/ListingsTabs"
 export function Listings() {
   return (
-    <div><ListingsTabs /></div>
-  )
+    <Container className="App">
+        <h2>Listings</h2>
+        <Form className="form">
+        <Col>
+          <ListingsTabs />
+        </Col>
+      </Form>
+    </Container>
+  );
 }
