+++ src/components/Listings/ListingsTable/ListingsTable/ListingsTableStyles.js
@@ -1,8 +1,5 @@
 import { makeStyles } from '@material-ui/core/styles'
 
-
-
-  
 export const tableStyles = makeStyles({
   table: {
     minWidth: 500,
