+++ src/components/Listings/ListingsTable/ListingsTablePagination/PaginationStyle.js
@@ -1,6 +1,5 @@
 import { makeStyles } from '@material-ui/core/styles'
 
-
 export const paginationStyles = makeStyles((theme) => ({
   root: {
     flexShrink: 0,
